+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Unreleased_
 
 Added:
 
+- `Add int64In/Out record support <../../pull/161>`_
 - `Enable setting alarm status of Out records <../../pull/157>`_
 - `Adding the non_interactive_ioc function <../../pull/156>`_